app/requirements.txt
fastapi
uvicorn==0.22.0
segment-geospatial
geopandas
shapely
fiona==1.9.4
pyproj==3.6.0
rtree==1.0.1
typing-extensions>=4.10.0
maplibre>=0.2.6
shiny>=0.7.0
rasterio
pillow